% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\docType{methods}
\name{add}
\alias{add}
\alias{add,NGScollation-method}
\alias{add,list-method}
\title{description of function add}
\usage{
\S4method{add}{NGScollation}(x, entry, stype = NULL, atype = NULL)

\S4method{add}{list}(x, entry, name = NULL, replace = T)
}
\arguments{
\item{x}{the NGScollation object}

\item{entry}{an object of package StefansExpressionSet}

\item{stype}{the type of the samples default=NULL}

\item{atype}{the type of the annoation default=NULL}

\item{name}{the name of the enw entry (optional)}

\item{replace}{allow replacement default T}

\item{x}{a list}

\item{entry}{the new entry}
}
\description{
Add a StefansExpressionSet to the NGScollation object

add an entry into a list
}

