% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findSample.R
\docType{methods}
\name{findSample}
\alias{findSample}
\alias{findSample,NGScollation-method}
\title{description of function findSample}
\usage{
findSample(x, sname)
}
\arguments{
\item{x}{the NGScollation}

\item{sname}{the sample name as defined in the sample tables}
}
\description{
searches all sample tables and reports the names of the datasets the sample is defined in.
}

