% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syncAnno.R
\docType{methods}
\name{syncAnno}
\alias{syncAnno}
\alias{syncAnno,NGScollation-method}
\title{description of function syncAnno}
\usage{
syncAnno(x)
}
\arguments{
\item{x}{the NGScollation object}
}
\description{
copy the putatives modified annotation and samples tables from the x object into the x$object(s)
}

